Clean up redirect logic to fix warnings and bugs.
